requirements.txt
# Core Dependencies
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional Dependencies
requests>=2.31.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
